fixed segmentation fault

+++ proj1/execute_U.c
@@ -6,5 +6,5 @@ void execute_U(){
     
     scanf("%hu\n", &user);
 
-    for(unsigned short a = 0; message_list[a].message[0] != '\0' && a < MAX_MSG; message_list[a].user_id == user ? printf("%s\n", message_list[a++].message) : a++);
+    for(unsigned int a = 0; message_list[a].message[0] != '\0' && a < MAX_MSG; message_list[a].user_id == user ? printf("%s\n", message_list[a++].message) : a++);
 }
